requirements.txt
# Dependencias principales verificadas para Python 3.13
fastapi==0.104.1
streamlit==1.28.1
uvicorn[standard]==0.35.0

# LangChain actualizado
langchain==0.3.27
langchain-community==0.3.27
langchain-openai==0.3.30
langchain-text-splitters==0.3.9
langchain-core==0.3.74

# Vector Store y Embeddings
chromadb==1.0.16
sentence-transformers==5.1.0

# Procesamiento PDF
PyPDF2==3.0.1
pdfplumber==0.10.3

# LLM Providers
openai==1.99.9
ollama==0.5.3

# Librerías ML
torch==2.8.0
transformers==4.55.2
numpy==2.3.2
scikit-learn==1.7.1

# Utilidades
python-multipart==0.0.20
python-dotenv==1.1.1
requests==2.32.4
aiofiles==24.1.0

# Testing y desarrollo
pytest==8.4.1
pytest-asyncio==1.1.0
black==25.1.0
flake8==7.3.0
